- Refactored code for new hierarchy
